@@ -30,15 +30,15 @@
   },
   "dependencies": {
     "@agm/core": "^3.0.0-beta.0",
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.7",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/common": "^18.2.7",
+    "@angular/compiler": "^18.2.7",
+    "@angular/core": "^18.2.7",
+    "@angular/forms": "^18.2.7",
+    "@angular/platform-browser": "^18.2.7",
+    "@angular/platform-browser-dynamic": "^18.2.7",
+    "@angular/router": "^18.2.7",
     "@asymmetrik/ngx-leaflet": "3.0.1",
     "@nebular/auth": "^7.0.5",
     "@nebular/eva-icons": "4.4.0",
@@ -79,10 +79,10 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
-    "@angular/cli": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/language-service": "17.3.12",
+    "@angular-devkit/build-angular": "^18.2.7",
+    "@angular/cli": "^18.2.7",
+    "@angular/compiler-cli": "^18.2.7",
+    "@angular/language-service": "18.2.7",
     "@compodoc/compodoc": "1.0.1",
     "@fortawesome/fontawesome-free": "^5.15.2",
     "@types/d3-color": "1.0.5",
